+++ jkube-kit/enricher/api/src/test/java/org/eclipse/jkube/kit/enricher/handler/JobHandlerTest.java
@@ -24,7 +24,7 @@ import org.eclipse.jkube.kit.config.resource.ResourceConfig;
 import org.eclipse.jkube.kit.config.resource.VolumeConfig;
 
 import io.fabric8.kubernetes.api.model.KubernetesListBuilder;
-import io.fabric8.kubernetes.api.model.batch.Job;
+import io.fabric8.kubernetes.api.model.batch.v1.Job;
 import mockit.Mocked;
 import org.junit.Before;
 import org.junit.Test;