+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Usage:
 ./scripts/extract-changelog-for-version.sh 1.3.37 5
 ```
 ### 1.4.0-SNAPSHOT
+* Fix #705: JIB assembly works on Windows
 
 ### 1.3.0
 * Fix #497: Assembly descriptor removed but still in documentation