+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Usage:
 ```
 ### 1.4.0-SNAPSHOT
 * Fix #705: JIB assembly works on Windows
+* Fix #714: feat: Helm support for Golang expressions
 
 ### 1.3.0
 * Fix #497: Assembly descriptor removed but still in documentation