+++ CHANGELOG.md
@@ -22,6 +22,8 @@ Usage:
 ```
 ### 1.4.0-SNAPSHOT
 * Fix #705: JIB assembly works on Windows
+* Fix #714: feat: Helm support for Golang expressions
+* Port fabric8io/docker-maven-plugin#1318: Update ECR autorization token URL
 * Fix #710: Support DockerImage as output for Openshift builds
 
 ### 1.3.0